Use Intersection Observer in Extension List
- Adds lazy loading to images
