+++ package.json
@@ -45,6 +45,7 @@
 		"axios": "^0.21.1",
 		"core-js": "2.6.4",
 		"element-ui": "^2.15.1",
+		"intersection-observer": "^0.12.0",
 		"iso-639-1": "^2.1.9",
 		"lodash.groupby": "^4.6.0",
 		"lodash.sortby": "^4.7.0",
