+++ src/extensions/README.md
@@ -8,4 +8,4 @@ lang: en-US
 
 List of available extensions to use with Tachiyomi, you can download them from here or from the app.
 
-<ExtensionList/>
+<Extensions/>
